# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Wedding",
        "description": "Pra casar API for wedding name Pra Casar",
        "termsOfService": "",
        "contact": {
            "email": "admin@approx.com.br"
        },
        "license": {
            "name": "Pro Produtores",
            "url": "https://www.proprodutores.com.br/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/invite/image/{image}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Show InviteImage by id",
                "description": "Show InviteImage by Id",
                "operationId": "App\\Http\\Controllers\\InviteImageController::getImage",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "InviteImage Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invite"
                ],
                "summary": "Update Image Url",
                "description": "If the User doesn't have a Invite, it will create a New Invite, or Update the existing Invite",
                "operationId": "App\\Http\\Controllers\\InviteController::update",
                "parameters": [
                    {
                        "name": "text",
                        "in": "path",
                        "description": "Invite's Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Invite attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Invite"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invite"
                ],
                "summary": "Update Invite Image",
                "description": "Update the InviteImage",
                "operationId": "App\\Http\\Controllers\\InviteImageController::update",
                "parameters": [
                    {
                        "name": "text",
                        "in": "path",
                        "description": "InviteImage Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "InviteImage attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InviteImage"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invite"
                ],
                "summary": "Delete Invite's Image",
                "description": "Delete Invite's Image by Invite's Id ",
                "operationId": "App\\Http\\Controllers\\InviteImageController::delete",
                "parameters": [
                    {
                        "name": "text",
                        "in": "path",
                        "description": "InviteImage Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/post": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create Post",
                "description": "Create a Post with a filter containing censured words chosen by the couple and using the image of the Guest who commented",
                "operationId": "App\\Http\\Controllers\\PostController::create",
                "requestBody": {
                    "description": "Post attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/{post}": {
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete Post",
                "description": "Delete Guest Post by User or if the Guest owns the post",
                "operationId": "App\\Http\\Controllers\\PostController::delete",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user/posts-for-aprove": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "List of for Aprove Posts",
                "description": "Return the list of Posts to Aprove",
                "operationId": "App\\Http\\Controllers\\PostController::forAprove",
                "requestBody": {
                    "description": "Post attributes",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "List Posts",
                "description": "Return the list of Posts to user or aproved posts to the user's guests",
                "operationId": "App\\Http\\Controllers\\PostController::index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/{post}/aprove": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Aprove the Post",
                "description": "Aprovall by the couple",
                "operationId": "App\\Http\\Controllers\\PostController::aprove",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Post attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostAprove"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/{post}/like": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Post Likes",
                "description": "Save the Post Likes on pivot table by post id",
                "operationId": "App\\Http\\Controllers\\PostController::postLike",
                "requestBody": {
                    "description": "Post attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostLike"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/{post}/likes": {
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Post Likes by Guests",
                "description": "Save the Post Likes on pivot table by post id",
                "operationId": "App\\Http\\Controllers\\PostController::likes",
                "requestBody": {
                    "description": "Post attributes",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuestLike"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/post/{post}/image": {
            "get": {
                "tags": [
                    "Invite"
                ],
                "summary": "Show Postimage by id",
                "description": "Show Postimage by Id",
                "operationId": "App\\Http\\Controllers\\PostController::getImage",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "Post Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GuestLike"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GuestLike": {
                "required": [
                    "guest_id",
                    "liked_id",
                    "liked"
                ],
                "properties": {
                    "guest_id": {
                        "type": "string",
                        "example": "JCM88976"
                    },
                    "liked_id": {
                        "type": "string",
                        "example": "JMF87676"
                    },
                    "liked": {
                        "type": "tinyint",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "PostLike": {
                "required": [
                    "post_id",
                    "guest_id"
                ],
                "properties": {
                    "post_id": {
                        "type": "integer",
                        "example": 5
                    },
                    "guest_id": {
                        "type": "string",
                        "example": "JMF87676"
                    }
                },
                "type": "object"
            },
            "PostAprove": {
                "required": [
                    "aproved"
                ],
                "properties": {
                    "aproved": {
                        "type": "tinyint",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "Invite": {
                "required": [
                    "id",
                    "user_id",
                    "bg_url"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "bg_url": {
                        "type": "string",
                        "example": "jiasudoiua.jpeg"
                    }
                },
                "type": "object"
            },
            "InviteImage": {
                "required": [
                    "id",
                    "image_url",
                    "width",
                    "height",
                    "x",
                    "y",
                    "layer",
                    "invite_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "image_url": {
                        "type": "string",
                        "example": "jaosidoqwoiklsa.jpeg"
                    },
                    "width": {
                        "type": "double",
                        "example": 1.41
                    },
                    "height": {
                        "type": "double",
                        "example": 850.06
                    },
                    "x": {
                        "type": "double",
                        "example": 25.69
                    },
                    "y": {
                        "type": "double",
                        "example": 103.56
                    },
                    "layer": {
                        "type": "integer",
                        "example": 10
                    },
                    "invite_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "Post": {
                "required": [
                    "id",
                    "text",
                    "guest_id",
                    "image_url",
                    "aproved"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "text": {
                        "type": "string",
                        "example": "Melhor Casal"
                    },
                    "guest_id": {
                        "type": "string",
                        "example": "TRU87856"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "post989820.png"
                    },
                    "aproved": {
                        "type": "tinyint",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}